ref(ngrx-ducks): move effect to ducks/
