+++ ngrx/ducks/src/lib/ducks/effect.ts
@@ -1,4 +1,4 @@
-import { ActionCreatorForEffect } from '../types';
+import { ActionCreatorForEffect } from '../typings';
 
 export function effect<T = void>(type: string): ActionCreatorForEffect<T> {
   const action = (payload: T) => ({
