+++ ngrx/ducks/src/lib/ducks/index.ts
@@ -1,1 +1,2 @@
 export * from './create-duck';
+export * from './effect';
