+++ ngrx/ducks/src/lib/typings/action-creator-for-effects.ts
