+++ ngrx/ducks/src/public_api.ts
@@ -1,11 +1,11 @@
 export * from './lib/core/create-reducer-from/create-reducer-from'; // deprecated
 export * from './lib/core/ducks/create-ducks'; // deprecated
-export * from './lib/core/ducks/effect';
 export * from './lib/core/types/ducks/ducks'; // deprecated
 export * from './lib/core/wired-actions/create-wired-action'; // deprecated
 export * from './lib/core/wired-actions/wire-up-actions'; // deprecated
 export * from './lib/decorators/action.decorator';
 export * from './lib/decorators/initial-state.decorator';
 export * from './lib/ducks/create-duck-service';
+export * from './lib/ducks/effect';
 export * from './lib/reducer/reducer-from';
 export * from './lib/typings/duck-service';
